+++ test/e2e/yarn-berry.test.ts
@@ -1,19 +1,44 @@
 import execa from "execa";
-import { ensureDir, readJson, writeJson } from "fs-extra";
+import {
+	copy,
+	createFile,
+	emptyDir,
+	ensureDir,
+	readJson,
+	writeJson,
+} from "fs-extra";
 import os from "os";
 import path from "path";
 import rimraf from "rimraf";
 import { promisify } from "util";
 import { packageManagers } from "../../src/index";
 
-describe("End to end tests - yarn", () => {
+describe("End to end tests - yarn berry", () => {
 	let testDir: string;
 
 	beforeEach(async () => {
 		// Create test directory
-		testDir = path.join(os.tmpdir(), "pak-test-yarn");
+		testDir = path.join(os.tmpdir(), "pak-test-yarn-berry");
 		await promisify(rimraf)(testDir);
 		await ensureDir(testDir);
+		// Upgrade it to yarn v3
+		for (const file of [".yarnrc.yml", ".yarn/releases/yarn-3.0.2.cjs"]) {
+			const source = path.join(__dirname, file);
+			const target = path.join(testDir, file);
+			await emptyDir(path.dirname(target));
+			await copy(source, target);
+		}
+		// Create empty yarn.lock, or it will look further up the tree
+		await createFile(path.join(testDir, "yarn.lock"));
+		// Keep yarn from failing when changing the lockfile
+		process.env.YARN_ENABLE_IMMUTABLE_INSTALLS = "0";
+	});
+
+	it("is actually using yarn berry", async () => {
+		jest.setTimeout(60000);
+		const yarn = new packageManagers.yarn();
+		yarn.cwd = testDir;
+		await expect(yarn.version()).resolves.toBe("3.0.2");
 	});
 
 	it("installs und uninstalls correctly", async () => {
@@ -61,12 +86,16 @@ describe("End to end tests - yarn", () => {
 		await writeJson(packageJsonPath, packageJson);
 
 		const yarn = new packageManagers.yarn();
+		yarn.loglevel = "verbose";
 		yarn.cwd = testDir;
 
 		// is-even@1.0.0 includes is-odd@^0.1.2, which also has newer versions (1.0.0+)
-		await yarn.install(["is-even@1.0.0"]);
-		let version = await execa.command(
-			'node -p require("is-odd/package.json").version',
+		let result = await yarn.install(["is-even@1.0.0"]);
+		expect(result.success).toBe(true);
+
+		let version = await execa(
+			"node",
+			["-p", 'require("is-odd/package.json").version'],
 			{
 				cwd: testDir,
 				reject: false,
@@ -75,13 +104,14 @@ describe("End to end tests - yarn", () => {
 		);
 		expect(version.stdout).toMatch(/^0\./);
 
-		const result = await yarn.overrideDependencies({
+		result = await yarn.overrideDependencies({
 			"is-odd": "1.0.0",
 		});
 		expect(result.success).toBe(true);
 
-		version = await execa.command(
-			'node -p require("is-odd/package.json").version',
+		version = await execa(
+			"node",
+			["-p", 'require("is-odd/package.json").version'],
 			{
 				cwd: testDir,
 				reject: false,
@@ -90,8 +120,9 @@ describe("End to end tests - yarn", () => {
 		);
 		expect(version.stdout).toBe("1.0.0");
 
-		version = await execa.command(
-			'node -p require("is-odd/package.json").version',
+		version = await execa(
+			"node",
+			["-p", 'require("is-odd/package.json").version'],
 			{
 				cwd: path.join(testDir, "node_modules/is-even"),
 				reject: false,
@@ -101,7 +132,8 @@ describe("End to end tests - yarn", () => {
 		expect(version.stdout).toBe("1.0.0");
 	});
 
-	it("does not install devDependencies, unless the environment is set to development", async () => {
+	// Yarn berry does not support the --production flag
+	it.skip("does not install devDependencies, unless the environment is set to development", async () => {
 		jest.setTimeout(60000);
 		const packageJson: Record<string, any> = {
 			name: "test",